+++ resources/views/components/table/td/reorder.blade.php
@@ -1,7 +1,14 @@
-@aware([ 'tableName', 'isTailwind', 'isBootstrap', 'isBootstrap4', 'isBootstrap5'])
+@aware([ 'tableName', 'isTailwind', 'isBootstrap', 'isBootstrap4', 'isBootstrap5', 'localisationPath'])
 @props(['rowID', 'rowIndex'])
 
-<x-livewire-tables::table.td.plain x-cloak x-show="currentlyReorderingStatus" wire:key="{{ $tableName }}-tbody-reorder-{{ $rowID }}" :displayMinimisedOnReorder="false">
+{{-- The drag handle is also keyboard-operable (#25): focus it (Tab) while
+     reordering and move the row with ArrowUp/ArrowDown. --}}
+<x-livewire-tables::table.td.plain x-cloak x-show="currentlyReorderingStatus" wire:key="{{ $tableName }}-tbody-reorder-{{ $rowID }}" :displayMinimisedOnReorder="false"
+    role="button"
+    aria-label="{{ __(($localisationPath ?? 'livewire-tables::core.').'Reorder') }}"
+    x-bind:tabindex="currentlyReorderingStatus ? 0 : -1"
+    x-on:keydown.arrow-up.prevent.stop="moveRow(event, -1)"
+    x-on:keydown.arrow-down.prevent.stop="moveRow(event, 1)">
     <svg
         x-cloak x-show="currentlyReorderingStatus"
         xmlns="http://www.w3.org/2000/svg"