+++ resources/views/components/tools/toolbar/items/column-select.blade.php
@@ -36,7 +36,7 @@
             </div>
 
             <div
-                x-cloak x-show="open"
+                x-cloak x-show="open" x-trap="open"
                 x-transition:enter="transition ease-out duration-100"
                 x-transition:enter-start="transform opacity-0 scale-95"
                 x-transition:enter-end="transform opacity-100 scale-100"
@@ -157,6 +157,7 @@
             </button>
 
             <div
+                x-trap="open"
                 x-bind:class="{ 'show': open }"
                 @class([$this->themeClasses('cs.menu')])
                 aria-labelledby="columnSelect-{{ $tableName }}"